JAVA: Fixed a number of imports that were broken from the previous refactor
