+++ src/main/java/com/synopsys/integration/alert/channel/jira/util/JiraMessageParser.java
@@ -32,7 +32,7 @@ import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
 
 import com.synopsys.integration.alert.channel.jira.model.IssueContentModel;
-import com.synopsys.integration.alert.common.channel.ChannelMessageParser;
+import com.synopsys.integration.alert.common.channel.message.ChannelMessageParser;
 import com.synopsys.integration.alert.common.channel.message.MessageSplitter;
 import com.synopsys.integration.alert.common.enumeration.ItemOperation;
 import com.synopsys.integration.alert.common.message.model.ComponentItem;
